Move satellite table to new file
